requirements.txt
anyio==4.10.0
beautifulsoup4==4.13.4
certifi==2025.8.3
charset-normalizer==3.4.3
deep-translator==1.11.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
langdetect==1.0.9
packaging==25.0
pluggy==1.6.0
pydub==0.25.1
Pygments==2.19.2
pytest==8.4.1
python-dotenv==1.1.1
python-telegram-bot==22.3
pytz==2025.2
requests==2.32.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SpeechRecognition==3.14.3
typing_extensions==4.14.1
urllib3==2.5.0
